Release v1.0.108

+++ src/main/webapp/scripts/dataacquisitionprojectmanagement/directives/metadata-type-state-card.directive.js
@@ -19,11 +19,13 @@ angular.module('metadatamanagementApp')
       replace: true,
       controllerAs: 'ctrl',
 
-      controller: function($scope, $rootScope) {
+      controller: function($scope, $rootScope, DataPackageIdBuilderService) {
         $scope.bowser = $rootScope.bowser;
         this.type = $scope.type;
         this.counts = $scope.counts;
         this.project = $scope.project;
+        this.dataPackageId =  DataPackageIdBuilderService
+          .buildDataPackageId(this.project.id);
 
         this.isAssignedDataProvider =
           ProjectUpdateAccessService.isAssignedToProject.bind(null,
@@ -101,12 +103,10 @@ angular.module('metadatamanagementApp')
             break;
           case 'publications':
             this.icon = 'assets/images/icons/related-publication.svg';
-            this.createState = 'publicationAssignment';
+            this.createState = '';
             this.searchState = 'related_publications';
             this.tooltip = 'search-management.buttons.' +
              'edit-publications-tooltip';
-            this.deleteTooltip = 'search-management.buttons.' +
-             'delete-publications-tooltip';
             break;
           case 'fake1':
             break;
@@ -185,9 +185,11 @@ angular.module('metadatamanagementApp')
           return ctrl.project.release;
         };
 
-        ctrl.delete = function() {
-          DeleteMetadataService.deleteAllOfType(ctrl.project, ctrl.type);
-        };
+        if (ctrl.type !== 'publications') {
+          ctrl.delete = function() {
+            DeleteMetadataService.deleteAllOfType(ctrl.project, ctrl.type);
+          };
+        }
 
         ctrl.edit = function(type) {
           if (ProjectUpdateAccessService.isUpdateAllowed(
